User Details
#20

OcipClient now exposes a getUserDetails method which returns a UserDetails object containing information about the logged in user. These details are also returned by the login method if called directly.

Along with this, a new configuration item has been added to Options to set the BroadWorks server version. There are two options available: 14sp4 (default) and 22. These affect what OCI method is called for logging in. Version 22 introduced a reseller feature, and setting the server version to 22 will set the resellerId on the UserDetails object.