TCP response fix
OCI-P over TCP ends its response with </BroadsoftDocument> followed by a newline. TCPTransport was not checking if the newline was receieved. In some situations, the newline might be still in the socket buffer and is read at the next request. This caused a newline to prefix the response which would cause an error when parsing it.