Fix for nillable abstract fields
Fixes #54

When a field's type is an abstract class and the value has been set to null, BroadWorks still expects type xsi:type attribute to be on the XML tag along with xsi:nil="true". This was not happening because the XML serializer does not know which concrete type to specify for the xsi:type property. It now will use the first concrete type it is aware of for the abstract type because it's not important which one is used since the value is null.